melhorei a organização
